+++ lib_i2s/src/i2s_frame_slave_impl.h
@@ -1,4 +1,4 @@
-// Copyright 2015-2023 XMOS LIMITED.
+// Copyright 2015-2024 XMOS LIMITED.
 // This Software is subject to the terms of the XMOS Public Licence: Version 1.
 #include <xs1.h>
 #include <xclib.h>
@@ -48,14 +48,17 @@ static void i2s_frame_slave0(client i2s_frame_callback_if i2s_i,
     // Since #pragma unsafe arrays is used need to ensure array won't overflow.
     assert((num_in << 1) <= 16);
 
+    i2s_config_t config;
+    config.slave_frame_synch_error = 0;
+
     if (num_data_bits == 32)
     {
         while(1){
             i2s_frame_slave_init_ports(p_dout, num_out, p_din, num_in, p_bclk, p_lrclk, bclk);
 
-            i2s_config_t config;
             i2s_restart_t restart = I2S_NO_RESTART;
             i2s_i.init(config, null);
+            config.slave_frame_synch_error = 0;
 
             //Get initial send data if output enabled
             if (num_out) i2s_i.send(num_out << 1, out_samps);
@@ -168,6 +171,11 @@ static void i2s_frame_slave0(client i2s_frame_callback_if i2s_i,
                 return;
             }
 
+            if(syncerror)
+            {
+                config.slave_frame_synch_error = 1;
+            }
+
         }// while(1)
     }
     else // else if num_data_bits != 32
@@ -181,6 +189,7 @@ static void i2s_frame_slave0(client i2s_frame_callback_if i2s_i,
             i2s_config_t config;
             i2s_restart_t restart = I2S_NO_RESTART;
             i2s_i.init(config, null);
+            config.slave_frame_synch_error = 0;
 
             // Get initial send data if output enabled
             if (num_out)
@@ -320,6 +329,11 @@ static void i2s_frame_slave0(client i2s_frame_callback_if i2s_i,
             if(restart == I2S_SHUTDOWN)
             {
                 return;
+            }
+
+            if(syncerror)
+            {
+                config.slave_frame_synch_error = 1;
             }
         }// while(1)
     } // if num_data_bits == 32